Agent/requirements.txt
livekit>=0.12.0
livekit-agents>=0.8.0
livekit-plugins-google>=0.6.0
python-dotenv>=1.0.0